requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Werkzeug==2.0.2
Flask-Script==2.0.6
Flask-Migrate==3.1.0
mysql-connector-python==8.0.26
python-dotenv==0.19.2
SQLAlchemy==1.4.32
gunicorn==20.1.0
Flask-Bcrypt==0.7.1
email-validator==1.1.3
pymysql==1.0.2
Flask-Bcrypt==0.7.1
Flask-WTF==0.15.1
prometheus-flask-exporter==0.18.4
prometheus_client
google-cloud-storage
uuid